fixed a crash that would occor on 64-bit devives caused by erroneous assumptions of type-sizes. And Fixed some static-analyizer warnings about losing precision due to erroneous assumptions about type sizes.

+++ Classes/MultiValueViewController.h
@@ -5,13 +5,13 @@
 @interface MultiValueViewController : UITableViewController {
 	id identifier;
 	NSArray *options, *images;
-	int selectedItemIndex;
+	NSUInteger selectedItemIndex;
 	UITableViewCell *selectedCell;
 }
 
 @property(strong) id identifier;
 @property(unsafe_unretained,assign) id<MultiValueViewControllerDelegate> delegate;
-@property(assign) int selectedItemIndex;
+@property(assign) NSUInteger selectedItemIndex;
 @property(copy) NSArray *options, *images;
 
 @end
@@ -19,6 +19,6 @@
 @protocol MultiValueViewControllerDelegate <NSObject>
 
 @required
-- (void) didSelectItemFromList: (MultiValueViewController *)multiValueViewController selectedItemIndex:(int)selectedItemIndex identifier:(id)identifier;
+- (void) didSelectItemFromList: (MultiValueViewController *)multiValueViewController selectedItemIndex:(NSUInteger)selectedItemIndex identifier:(id)identifier;
 
 @end
